src/main/webapp/app/entities/consumer/shop-two/route/shop-two-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IShopTwo, ShopTwo } from '../shop-two.model';
import { ShopTwoService } from '../service/shop-two.service';

import { ShopTwoRoutingResolveService } from './shop-two-routing-resolve.service';

describe('ShopTwo routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: ShopTwoRoutingResolveService;
  let service: ShopTwoService;
  let resultShopTwo: IShopTwo | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(ShopTwoRoutingResolveService);
    service = TestBed.inject(ShopTwoService);
    resultShopTwo = undefined;
  });

  describe('resolve', () => {
    it('should return IShopTwo returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultShopTwo = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultShopTwo).toEqual({ id: 123 });
    });

    it('should return new IShopTwo if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultShopTwo = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultShopTwo).toEqual(new ShopTwo());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as ShopTwo })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultShopTwo = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultShopTwo).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
